Conflicts:
	data/Entity.php
	data/source/mongo_db/Exporter.php
	storage/session/adapter/Php.php
	tests/cases/data/CollectionTest.php

+++ console/command/create/View.php
@@ -2,7 +2,7 @@
 /**
  * Lithium: the most rad php framework
  *
- * @copyright     Copyright 2011, Union of RAD (http://union-of-rad.org)
+ * @copyright     Copyright 2012, Union of RAD (http://union-of-rad.org)
  * @license       http://opensource.org/licenses/bsd-license.php The BSD License
  */
 
@@ -44,7 +44,13 @@ class View extends \lithium\console\command\Create {
 				}
 			}
 			$directory = str_replace($this->_library['path'] . '/', '', $directory);
-
+			if (file_exists($file)) {
+				$prompt = "{$file} already exists. Overwrite?";
+				$choices = array('y', 'n');
+				if ($this->in($prompt, compact('choices')) != 'y') {
+					return "{$params['file']} skipped.";
+				}
+			}
 			if (file_put_contents($file, "<?php\n\n{$result}\n\n?>")) {
 				return "{$params['file']}.php created in {$directory}.";
 			}
